// EscalafonFrontend/src/app/interfaces/formacion-academica.ts
export interface FormacionAcademica {
    /**
     * @type { number } - iFormAcadId
     */
    iFormAcadId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iEduNivEdId
     */
    iEduNivEdId?: number;

    /**
     * @type { number } - iEduTipoSecId
     */
    iEduTipoSecId?: number;

    /**
     * @type { number } - iPaisId
     */
    iPaisId?: number;

    /**
     * @type { number } - iEduTipSupId
     */
    iEduTipSupId?: number;

    /**
     * @type { string } - cFormAcadCiudad
     */
    cFormAcadCiudad?: string;

    /**
     * @type { number } - iDptoId
     */
    iDptoId?: number;

    /**
     * @type { number } - iPrvnId
     */
    iPrvnId?: number;

    /**
     * @type { number } - iDsttId
     */
    iDsttId?: number;

    /**
     * @type { number } - iEduSitAcadId
     */
    iEduSitAcadId?: number;

    /**
     * @type { number } - iEduNivEspId
     */
    iEduNivEspId?: number;

    /**
     * @type { number } - iEduCarrId
     */
    iEduCarrId?: number;

    /**
     * @type { number } - iEduProGenId
     */
    iEduProGenId?: number;

    /**
     * @type { number } - iEduProProfId
     */
    iEduProProfId?: number;

    /**
     * @type { number } - iEduProId
     */
    iEduProId?: number;

    /**
     * @type { number } - iEduGradAlcId
     */
    iEduGradAlcId?: number;

    /**
     * @type { number } - iEduSitAcadAnioInicio
     */
    iEduSitAcadAnioInicio?: number;

    /**
     * @type { number } - iEduSitAcadAnioFin
     */
    iEduSitAcadAnioFin?: number;

    /**
     * @type { string } - cEduSitAcadCentroEstudio
     */
    cEduSitAcadCentroEstudio?: string;

    /**
     * @type { string } - cFormAcadTitulo
     */
    cFormAcadTitulo?: string;

    /**
     * @type { string } - cFormAcadEspecialidad
     */
    cFormAcadEspecialidad?: string;

    /**
     * @type { Date } - dtFormAcadFechaExpedicion
     */
    dtFormAcadFechaExpedicion?: Date;

    /**
     * @type { number } - iEduTipCentId
     */
    iEduTipCentId?: number;

    /**
     * @type { number } - iEduCentRegId
     */
    iEduCentRegId?: number;

    /**
     * @type { string } - cFormAcadCentroRegistro
     */
    cFormAcadCentroRegistro?: string;

    /**
     * @type { string } - cFormAcadNumeroRegistro
     */
    cFormAcadNumeroRegistro?: string;

    /**
     * @type { Date } - dtFormAcadFechaRegistro
     */
    dtFormAcadFechaRegistro?: Date;

    /**
     * @type { Date } - dtFormAcadFechaDocumento
     */
    dtFormAcadFechaDocumento?: Date;

    /**
     * @type { string } - cFormAcadNumeroDocumento
     */
    cFormAcadNumeroDocumento?: string;

    /**
     * @type { string } - cFormAcadCentroEstudios
     */
    cFormAcadCentroEstudios?: string;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { string } - cFormAcadEspecialidadNivel
     */
    cFormAcadEspecialidadNivel?: string;

}