// EscalafonFrontend/src/app/interfaces/pais.ts
export interface Pais {
    /**
     * @type { number } - iPaisId
     */
    iPaisId?: number;

    /**
     * @type { string } - cPaisCodigo
     */
    cPaisCodigo?: string;

    /**
     * @type { string } - cPaisNombre
     */
    cPaisNombre?: string;

    /**
     * @type { string } - cPaisAbreviatura1
     */
    cPaisAbreviatura1?: string;

    /**
     * @type { string } - cPaisAbreviatura2
     */
    cPaisAbreviatura2?: string;

    /**
     * @type { string } - cPaisCodigoSunat
     */
    cPaisCodigoSunat?: string;

    /**
     * @type { string } - cPaisGentilicio
     */
    cPaisGentilicio?: string;

}