// EscalafonFrontend/src/app/interfaces/departamento.ts
export interface Departamento {
    /**
     * @type { number } - iDptoId
     */
    iDptoId?: number;

    /**
     * @type { number } - iPaisId
     */
    iPaisId?: number;

    /**
     * @type { string } - cDptoCodigo
     */
    cDptoCodigo?: string;

    /**
     * @type { string } - cDptoNombre
     */
    cDptoNombre?: string;

    /**
     * @type { string } - cDptoAbreviatura
     */
    cDptoAbreviatura?: string;

    /**
     * @type { number } - iDptoEstado
     */
    iDptoEstado?: number;

    /**
     * @type { string } - cDptoGentilicio
     */
    cDptoGentilicio?: string;

}